refactor: is<Type>함수 typeChecker로 통합
