+++ frontend/src/utils/index.ts
@@ -2,4 +2,4 @@ export { default as getRandomInt } from '@utils/getRandomInt';
 export { default as noop } from '@utils/noop';
 export { mqBetween, mqDown, mqUp } from '@utils/media-query';
 export { changeDateSeperator, getNextYear, getToday, yyyymmddTommdd } from '@utils/dates';
-export { isObject, isString } from '@utils/type-checks';
+export { isFunction, isNullOrUndefined, isNumber, isObject, isString } from '@utils/typeChecker';
