--- frontend/src/utils/isNullOrUndefined.ts
@@ -1,3 +1,0 @@
-const isNullOrUndefined = (value: unknown): value is null | undefined => value === undefined || value === null;
-
-export default isNullOrUndefined;
